moved prepare_blender_object to new blender_utils.py file
